+++ mobile/android/android-components/components/service/pocket/src/main/java/mozilla/components/service/pocket/stories/db/PocketRecommendationsDatabase.kt
@@ -23,7 +23,7 @@ import mozilla.components.service.pocket.spocs.db.SpocsDao
         SpocEntity::class,
         SpocImpressionEntity::class
     ],
-    version = 3
+    version = 4,
 )
 internal abstract class PocketRecommendationsDatabase : RoomDatabase() {
     abstract fun pocketRecommendationsDao(): PocketRecommendationsDao
@@ -51,6 +51,7 @@ internal abstract class PocketRecommendationsDatabase : RoomDatabase() {
                     Migrations.migration_1_2,
                     Migrations.migration_2_3,
                     Migrations.migration_1_3,
+                    Migrations.migration_3_4,
                 )
                 .build().also {
                     instance = it
@@ -99,6 +100,54 @@ internal object Migrations {
     val migration_1_3 = object : Migration(1, 3) {
         override fun migrate(database: SupportSQLiteDatabase) {
             database.createNewSpocsTables()
+        }
+    }
+
+    /**
+     * Migration for when adding a new index to the spoc impression entity.
+     */
+    val migration_3_4 = object : Migration(3, 4) {
+        override fun migrate(database: SupportSQLiteDatabase) {
+            // Rename the old tables to allow creating new ones
+            database.execSQL(
+                "ALTER TABLE `${PocketRecommendationsDatabase.TABLE_NAME_SPOCS}` " +
+                    "RENAME TO temp_spocs"
+            )
+            database.execSQL(
+                "ALTER TABLE `${PocketRecommendationsDatabase.TABLE_NAME_SPOCS_IMPRESSIONS}` " +
+                    "RENAME TO temp_spocs_impressions"
+            )
+
+            // Create new tables with the new schema
+            database.createNewSpocsTables()
+            database.execSQL(
+                "CREATE INDEX IF NOT EXISTS `index_spocs_impressions_spocId` " +
+                    "ON `${PocketRecommendationsDatabase.TABLE_NAME_SPOCS_IMPRESSIONS}` (`spocId`)"
+            )
+
+            // Copy the old data to the new tables
+            database.execSQL(
+                "INSERT INTO " +
+                    "'${PocketRecommendationsDatabase.TABLE_NAME_SPOCS}' (" +
+                    "id, url, title, imageUrl, sponsor, clickShim, impressionShim, " +
+                    "priority, lifetimeCapCount, flightCapCount, flightCapPeriod" +
+                    ") SELECT " +
+                    "id, url, title, imageUrl, sponsor, clickShim, impressionShim, " +
+                    "priority, lifetimeCapCount, flightCapCount, flightCapPeriod " +
+                    "FROM temp_spocs"
+            )
+            database.execSQL(
+                "INSERT INTO " +
+                    "'${PocketRecommendationsDatabase.TABLE_NAME_SPOCS_IMPRESSIONS}' (" +
+                    "spocId, impressionId, impressionDateInSeconds" +
+                    ") SELECT " +
+                    "spocId, impressionId, impressionDateInSeconds " +
+                    "FROM temp_spocs_impressions"
+            )
+
+            // Cleanup
+            database.execSQL("DROP TABLE temp_spocs")
+            database.execSQL("DROP TABLE temp_spocs_impressions")
         }
     }